+++ .codespellrc
@@ -1,0 +1,2 @@
+[codespell]
+ignore-words-list = crate
