+++ Cargo.toml
@@ -19,7 +19,7 @@ blake3 = "1"
 bytes = "1"
 bytesize = "1"
 chrono = "0.4"
-clap = "3"
+clap = { version = "4", features = ["env", "string"] }
 colored = "2"
 crossbeam = "0.8"
 dirs = "5"
