+++ src/cli/commands/cmd_mb.rs
@@ -1,0 +1,29 @@
+use clap::{Arg, Command};
+
+pub fn subcommand_mb() -> Command {
+    Command::new("mb")
+        .about("Make a bucket")
+        .arg(
+            Arg::new("arguments")
+                .help("<s3 provider>/<bucket name>")
+                .required(true)
+                .num_args(1),
+        )
+        .arg(
+            Arg::new("acl")
+                .help("The canned ACL to apply to the object example")
+                .long("acl")
+                .value_parser([
+                    "private",
+                    "public-read",
+                    "public-read-write",
+                    "authenticated-read",
+                    "aws-exec-read",
+                    "bucket-owner-read",
+                    "bucket-owner-full-control",
+                ])
+                .default_value("private")
+                .short('a')
+                .num_args(1),
+        )
+}
