+++ src/cli/dispatch.rs
@@ -1,8 +1,7 @@
-use crate::s3m::start::Action;
+use crate::cli::actions::Action;
 use anyhow::{anyhow, Context, Result};
 use colored::Colorize;
-use std::collections::BTreeMap;
-use std::path::PathBuf;
+use std::{collections::BTreeMap, path::PathBuf};
 
 // return Action based on the command or subcommand
 pub fn dispatch(
@@ -36,11 +35,7 @@ pub fn dispatch(
         Some("acl") => {
             let key = hbp_empty(hbp)?;
             let sub_m = sub_m("acl")?;
-            let acl = if sub_m.is_present("acl") {
-                Some(sub_m.value_of("acl").unwrap_or_default().to_string())
-            } else {
-                None
-            };
+            let acl = sub_m.get_one("acl").map(|s: &String| s.to_string());
             Ok(Action::ACL { key, acl })
         }
 
@@ -48,8 +43,13 @@ pub fn dispatch(
         Some("get") => {
             let key = hbp_empty(hbp)?;
             let sub_m = sub_m("get")?;
-            let get_head = sub_m.is_present("HeadObject");
-            let args: Vec<&str> = sub_m.values_of("arguments").unwrap_or_default().collect();
+            let get_head = sub_m.get_one("HeadObject").copied().unwrap_or(false);
+            let args: Vec<&str> = sub_m
+                .get_many::<String>("arguments")
+                .unwrap_or_default()
+                .map(|s| s.as_str())
+                .collect();
+
             let dest = if args.len() == 2 {
                 Some(args[1].to_string())
             } else {
@@ -59,31 +59,21 @@ pub fn dispatch(
                 key,
                 get_head,
                 dest,
-                quiet: sub_m.is_present("quiet"),
+                quiet: sub_m.contains_id("quiet"),
             })
         }
 
         // ListObjects
         Some("ls") => {
             let sub_m = sub_m("ls")?;
-            let prefix = if sub_m.is_present("prefix") {
-                Some(sub_m.value_of("prefix").unwrap_or_default().to_string())
-            } else {
-                None
-            };
-            let start_after = if sub_m.is_present("start-after") {
-                Some(
-                    sub_m
-                        .value_of("start-after")
-                        .unwrap_or_default()
-                        .to_string(),
-                )
-            } else {
-                None
-            };
+            let prefix = sub_m.get_one("prefix").map(|s: &String| s.to_string());
+            let start_after = sub_m.get_one("start-after").map(|s: &String| s.to_string());
             Ok(Action::ListObjects {
                 bucket,
-                list_multipart_uploads: sub_m.is_present("ListMultipartUploads"),
+                list_multipart_uploads: sub_m
+                    .get_one("ListMultipartUploads")
+                    .copied()
+                    .unwrap_or(false),
                 prefix,
                 start_after,
             })
@@ -93,11 +83,9 @@ pub fn dispatch(
         Some("mb") => match bucket {
             Some(_) => {
                 let sub_m = sub_m("mb")?;
-                let acl = if sub_m.is_present("acl") {
-                    sub_m.value_of("acl").unwrap_or_default().to_string()
-                } else {
-                    String::from("private")
-                };
+                let acl = sub_m
+                    .get_one("acl")
+                    .map_or_else(|| String::from("private"), |s: &String| s.to_string());
                 Ok(Action::MakeBucket { acl })
             }
             None => Err(anyhow!("Bucket name missing, <s3 provider>/<bucket>")),
@@ -107,7 +95,9 @@ pub fn dispatch(
         Some("rm") => {
             let key = hbp_empty(hbp)?;
             let sub_m = sub_m("rm")?;
-            let upload_id = sub_m.value_of("UploadId").unwrap_or_default().to_string();
+            let upload_id = sub_m
+                .get_one("UploadId")
+                .map_or_else(String::new, |s: &String| s.to_string());
             Ok(Action::DeleteObject { key, upload_id })
         }
 
@@ -115,7 +105,7 @@ pub fn dispatch(
         Some("share") => {
             let key = hbp_empty(hbp)?;
             let sub_m = sub_m("share")?;
-            let expire = sub_m.value_of("expire").unwrap().parse::<usize>()?;
+            let expire = sub_m.get_one::<usize>("expire").map_or_else(|| 0, |s| *s);
             Ok(Action::ShareObject { key, expire })
         }
 
@@ -123,21 +113,27 @@ pub fn dispatch(
         _ => {
             let key = hbp_empty(hbp)?;
             let mut src: Option<String> = None;
-            let args: Vec<&str> = matches.values_of("arguments").unwrap_or_default().collect();
+            let args: Vec<&str> = matches
+                .get_many::<String>("arguments")
+                .unwrap_or_default()
+                .map(|s| s.as_str())
+                .collect();
             if args.len() == 2 {
                 src = Some(args[0].to_string());
             }
-            let acl = if matches.is_present("acl") {
-                Some(matches.value_of("acl").unwrap_or_default().to_string())
-            } else {
-                None
-            };
-            let meta = if matches.is_present("meta") {
+
+            let acl = matches.get_one("acl").map(|s: &String| s.to_string());
+
+            let meta = if matches
+                .get_one("meta")
+                .map(|s: &String| s.to_string())
+                .is_some()
+            {
                 Some(
                     matches
-                        .value_of("meta")
+                        .get_one("meta")
+                        .map(|s: &String| s.to_string())
                         .unwrap_or_default()
-                        .to_string()
                         .split(';')
                         .map(|s| s.split_once('=').unwrap())
                         .map(|(key, val)| {
@@ -155,8 +151,8 @@ pub fn dispatch(
                 file: src,
                 s3m_dir,
                 key,
-                pipe: matches.is_present("pipe"),
-                quiet: matches.is_present("quiet"),
+                pipe: matches.get_one("pipe").copied().unwrap_or(false),
+                quiet: matches.get_one("quiet").copied().unwrap_or(false),
             })
         }
     }
