+++ src/lib.rs
@@ -1,3 +1,3 @@
+pub mod cli;
 pub mod options;
 pub mod s3;
-pub mod s3m;
