--- src/s3m/args/mod.rs
@@ -1,3 +1,0 @@
-pub mod command;
-pub mod subcommands;
-pub mod validators;
